refactor: convert to typescript
