+++ components/date.tsx
@@ -1,0 +1,6 @@
+import { parseISO, format } from 'date-fns'
+
+export default function Date({ dateString }: { dateString: string }) {
+    const date = parseISO(dateString)
+    return <time dateTime={dateString}> {format(date, 'yyyy-MM-dd HH:mm')} </time>
+}
