+++ components/fontawesome.tsx
@@ -10,6 +10,5 @@ export default function Fa({
 }: any) {
     config.autoAddCss = false;
     library.add(faGithub, faKeybase, faSpeakerDeck, faTwitter, faInstagram, faRssSquare, faTags, faExclamationTriangle)
-    // @ts-expect-error ts-migrate(2749) FIXME: 'FontAwesomeIcon' refers to a value, but is being ... Remove this comment to see the full error message
     return <FontAwesomeIcon icon={[iconPrefix, iconName]} title={title} size='xs' />
 }
