+++ pages/_document.tsx
@@ -1,0 +1,43 @@
+import Document, { Html, Head, Main, NextScript } from 'next/document'
+
+import { GA_TRACKING_ID } from '../lib/gtag'
+
+class MyDocument extends Document {
+    render() {
+        return (
+            <Html lang="ja">
+                <Head>
+                    {/* Global Site Tag (gtag.js) - Google Analytics */}
+                    <script
+                        async
+                        src={`https://www.googletagmanager.com/gtag/js?id=${GA_TRACKING_ID}`}
+                    />
+                    <script
+                        dangerouslySetInnerHTML={{
+                            __html: `
+                    window.dataLayer = window.dataLayer || [];
+                    function gtag(){dataLayer.push(arguments);}
+                    gtag('js', new Date());
+                    gtag('config', '${GA_TRACKING_ID}', {
+                    page_path: window.location.pathname,
+                    });
+                `,
+                        }}
+                    />
+                    <meta name="application-name" content="the world as code" />
+                    <meta name="apple-mobile-web-app-capable" content="yes" />
+                    <meta name="apple-mobile-web-app-status-bar-style" content="default" />
+                    <meta name="apple-mobile-web-app-title" content="the world as code" />
+                    <meta name="mobile-web-app-capable" content="yes" />
+                    <link rel="manifest" href="/manifest.json" />
+                </Head>
+                <body className="line-numbers">
+                    <Main />
+                    <NextScript />
+                </body>
+            </Html>
+        )
+    }
+}
+
+export default MyDocument
